Bank Integration
Added a pair of Wallet classes that make a bank transfer and give the server a receipt. Added a pair of Auth classes that charge a fee. Modified Client to async-ly get the payment (since communicating with the bank is asynchronous) [not strictly necessary but seemed reasonable]. Created the new bank_functional1.py test script.